@@ -1,7 +1,7 @@
 package com.example.pushnotificationsintegrationtests
 
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.pusher.pushnotifications.PushNotificationsInstance
 import com.pusher.pushnotifications.auth.TokenProvider
 import com.pusher.pushnotifications.fcm.MessagingService
@@ -35,11 +35,11 @@ import java.util.concurrent.TimeUnit
 @RunWith(AndroidJUnit4::class)
 class ClearAllStateTest {
   val instanceId = "00000000-1241-08e9-b379-377c32cd1e81"
-  val context = InstrumentationRegistry.getTargetContext()
+  val context = InstrumentationRegistry.getInstrumentation().targetContext
   val errolClient = ErrolAPI(instanceId, "http://localhost:8080")
 
   fun getStoredDeviceId(): String? {
-    val deviceStateStore = InstanceDeviceStateStore(InstrumentationRegistry.getTargetContext(), instanceId)
+    val deviceStateStore = InstanceDeviceStateStore(InstrumentationRegistry.getInstrumentation().targetContext, instanceId)
     return deviceStateStore.deviceId
   }
 
@@ -56,7 +56,7 @@ class ClearAllStateTest {
   @Before
   @After
   fun wipeLocalState() {
-    val deviceStateStore = InstanceDeviceStateStore(InstrumentationRegistry.getTargetContext(), instanceId)
+    val deviceStateStore = InstanceDeviceStateStore(InstrumentationRegistry.getInstrumentation().targetContext, instanceId)
 
     await.atMost(1, TimeUnit.SECONDS) until {
       assertTrue(deviceStateStore.clear())
