@@ -1,7 +1,7 @@
 package com.example.pushnotificationsintegrationtests.internal
 
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.pusher.pushnotifications.internal.InstanceDeviceStateStore
 import org.awaitility.kotlin.await
 import org.awaitility.kotlin.until
@@ -21,7 +21,7 @@ import java.util.concurrent.TimeUnit
  */
 @RunWith(AndroidJUnit4::class)
 class InstanceDeviceStateStoreTest {
-  val context = InstrumentationRegistry.getTargetContext()
+  val context = InstrumentationRegistry.getInstrumentation().targetContext
   val instanceId1 = "00000000-1241-08e9-b379-377c32cd1e80"
   val instanceId2 = "00000000-1241-08e9-b379-377c32cd1e81"
 
