@@ -4,10 +4,9 @@ import android.app.Activity
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import com.firebase.jobdispatcher.*
+import androidx.work.*
 import com.google.gson.Gson
 import com.google.gson.JsonSyntaxException
-import com.pusher.pushnotifications.internal.DeviceStateStore
 import com.pusher.pushnotifications.internal.InstanceDeviceStateStore
 import com.pusher.pushnotifications.logging.Logger
 import com.pusher.pushnotifications.reporting.api.OpenEvent
@@ -28,7 +27,7 @@ class OpenNotificationActivity: Activity() {
             i = Intent()
             i.action = clickAction
         } else {
-            i = packageManager.getLaunchIntentForPackage(packageName)
+            i = packageManager.getLaunchIntentForPackage(packageName)!!
         }
         
         i.replaceExtras(bundle)
@@ -79,17 +78,17 @@ class OpenNotificationActivity: Activity() {
                    timestampSecs = System.currentTimeMillis() / 1000
                 )
 
-                val dispatcher = FirebaseJobDispatcher(GooglePlayDriver(applicationContext))
-                val job = dispatcher.newJobBuilder()
-                        .setService(ReportingJobService::class.java)
-                        .setTag("pusher.open.publishId=${pusherData.publishId}")
-                        .setConstraints(Constraint.ON_ANY_NETWORK)
-                        .setRetryStrategy(RetryStrategy.DEFAULT_EXPONENTIAL)
-                        .setLifetime(Lifetime.UNTIL_NEXT_BOOT)
-                        .setExtras(ReportingJobService.toBundle(reportEvent))
+                val reportWorker = OneTimeWorkRequest.Builder(ReportingWorker::class.java)
+                        .setConstraints(Constraints.Builder()
+                                .setRequiredNetworkType(NetworkType.CONNECTED)
+                                .build())
+                        .setInputData(ReportingWorker.toInputData(reportEvent))
                         .build()
 
-                dispatcher.mustSchedule(job)
+                val workManagerInstance = WorkManager.getInstance(applicationContext)
+                workManagerInstance.enqueueUniqueWork("pusher.open.publishId=${pusherData.publishId}",
+                        ExistingWorkPolicy.KEEP,
+                        reportWorker)
 
                 startIntent(intent.extras, pusherData.clickAction)
             } catch (_: JsonSyntaxException) {
